refactor: per-component subfolders + seo module barrel
Tightens repo structure without changing behavior:

- Move 13 flat components into per-component subfolders with index barrels,
  matching the existing pattern (home/Applications, ui/Button, etc.).
  Components affected: SectionHead, AccessoriesShell, AccessoriesSideNav,
  CompanyShell, CompanySideNav, Turnstile, LogoLockup, Logomark, PageShell,
  SearchPanel, SearchTriggerButton, ProductThumb, DocRow.
- Add missing index.ts barrel to products/SectionHeader/.
- Resolve src/lib/seo.ts <-> src/lib/seo/ shadow: move seo.ts -> seo/metadata.ts
  and add seo/index.ts re-export. All existing import sites continue to resolve.
- Shorten 5 deep imports of products/SectionHeader/SectionHeader to use the
  new barrel.
- Co-locate component CSS files into their new subfolders.
