+++ src/components/accessories/AccessoriesShell/index.ts
@@ -1,0 +1,1 @@
+export { AccessoriesShell } from "./AccessoriesShell";
