+++ src/components/accessories/AccessoriesSideNav/index.ts
@@ -1,0 +1,1 @@
+export { AccessoriesSideNav } from "./AccessoriesSideNav";
