+++ src/components/company/CompanyShell/index.ts
@@ -1,0 +1,1 @@
+export { CompanyShell } from "./CompanyShell";
