+++ src/components/company/CompanySideNav/index.ts
@@ -1,0 +1,1 @@
+export { CompanySideNav } from "./CompanySideNav";
