+++ src/components/home/SectionHead/SectionHead.tsx
