+++ src/components/home/SectionHead/index.ts
@@ -1,0 +1,1 @@
+export { SectionHead } from "./SectionHead";
