+++ src/components/layout/Logomark/Logomark.tsx
