+++ src/components/layout/PageShell/index.ts
@@ -1,0 +1,1 @@
+export { PageShell } from "./PageShell";
