+++ src/components/layout/Search/SearchTriggerButton/index.ts
@@ -1,0 +1,1 @@
+export { SearchTriggerButton } from "./SearchTriggerButton";
