+++ src/components/products/ProductThumb/ProductThumb.tsx
