+++ src/components/products/ProductThumb/index.ts
@@ -1,0 +1,1 @@
+export { ProductThumb } from "./ProductThumb";
