+++ src/components/resources/DocRow/index.ts
@@ -1,0 +1,1 @@
+export { DocRow } from "./DocRow";
